refactor(backend): replace old sandbox with new sandbox
